fix(deepseek-v4): cast attn_sink to queries.dtype to fix bfloat16 SDPA crash
MLXFast.scaledDotProductAttention requires sinks to promote to the output
dtype (bfloat16). attn_sink is initialized as float32 zeros and may be
loaded from the checkpoint as float32, causing the fatal error:
  'Type of sinks must promote to output type bfloat16'

Fix: add .asType(queries.dtype) cast at the call site before passing
sinksToUse to deepseekAttentionWithSinks().

Also adds performance benchmark results (M5 Pro 64GB) to README.

+++ Libraries/MLXLLM/Models/DeepseekV4.swift
@@ -443,7 +443,12 @@ class DeepseekV4Attention: Module {
         // --- Attention ---
         // Pass kFull as both keys and values; cache update happens inside.
         // Apply attn_sink (per-head bias) to attention logits when non-zero.
-        let sinksToUse: MLXArray? = attn_sink.sum().item(Float.self) != 0 ? attn_sink : nil
+        // Cast to queries.dtype (bfloat16) — attn_sink may be loaded as float32 from the
+        // checkpoint, but MLXFast.scaledDotProductAttention requires sinks to promote to
+        // the output dtype (bfloat16); float32 does not satisfy this constraint.
+        let sinksToUse: MLXArray? = attn_sink.sum().item(Float.self) != 0
+            ? attn_sink.asType(queries.dtype)
+            : nil
         let output = deepseekAttentionWithSinks(
             queries: queries,
             keys: kFull,
